Break source_dist.rs into a module
