+++ crates/puffin-distribution/src/lib.rs
@@ -2,7 +2,7 @@ pub use distribution_database::{DistributionDatabase, DistributionDatabaseError}
 pub use download::{BuiltWheel, DiskWheel, InMemoryWheel, LocalWheel};
 pub use index::{BuiltWheelIndex, RegistryWheelIndex};
 pub use reporter::Reporter;
-pub use source_dist::{SourceDistCachedBuilder, SourceDistError};
+pub use source::{SourceDistCachedBuilder, SourceDistError};
 pub use unzip::Unzip;
 
 mod distribution_database;
@@ -11,5 +11,5 @@ mod error;
 mod index;
 mod locks;
 mod reporter;
-mod source_dist;
+mod source;
 mod unzip;
